Update pom.xml and add checkstyle verification. (#38)
* Adding .editorconfig.

* Format pom.xml

* Update build tool versions.

* Adding pluginsManagement element.

* Fix compiler errors for rawtypes.

* Update ci.yml.

* Use java11 as default profile.

* Fix checkstyle issues.

* Updating documentation and config for javadocs.

+++ src/test/java/com/microsoft/azure/proton/transport/ws/impl/WebSocketImplTest.java
@@ -1085,7 +1085,7 @@ public class WebSocketImplTest {
         random.nextBytes(data);
 
         byte finbit = (byte) (WebSocketHeader.FINBIT_MASK & 0xFF);
-        byte opcode = (byte) WebSocketHeader.OPCODE_MASK & 0x2;
+        byte opcode = WebSocketHeader.OPCODE_MASK & 0x2;
         byte firstbyte = (byte) (finbit | opcode);
         byte secondbyte = (byte) (size - 2);
 
@@ -1268,7 +1268,7 @@ public class WebSocketImplTest {
         byte[] message = new byte[chunkSize];
 
         byte finbit = (byte) (WebSocketHeader.FINBIT_MASK & 0xFF);
-        byte opcode = (byte) WebSocketHeader.OPCODE_MASK & 0x2;
+        byte opcode = WebSocketHeader.OPCODE_MASK & 0x2;
         byte firstbyte = (byte) (finbit | opcode);
         byte secondbyte = (byte) payloadLength;
